requirements.txt
flask
pillow
qrcode
yt-dlp
gunicorn
